#33 refactor sending out emails
