+++ Libraries/DataLayer/RabbitMQ/CreditCardInformationChangedQueueSubscribe.cs
@@ -4,7 +4,7 @@ using System;
 using System.Net.Mail;
 using DataLayer.RabbitMQ.QueueMessages;
 using System.Text.Json;
-using System.Net;
+using Common.Email;
 
 namespace DataLayer.RabbitMQ
 {
@@ -31,22 +31,13 @@ namespace DataLayer.RabbitMQ
             CreditCardInformationChangedQueueMessage message = JsonSerializer.Deserialize<CreditCardInformationChangedQueueMessage>(e.Body.ToArray());
             SmtpClient SmtpServer = new SmtpClient("smtp.gmail.com");
             SmtpServer.Port = 587;
-            using (MailMessage mail = new MailMessage())
-            {
-                mail.From = new MailAddress("support@encryptionapiservices.com");
-                mail.To.Add(message.UserEmail);
-                mail.Subject = "Credit Card Changed - Encryption API Services";
-                mail.Body = "We noticed that you changed your credit card information recently. If this wasn't you we recommend changing your password " + String.Format("<a href='" + Environment.GetEnvironmentVariable("Domain") + "/#/forgot-password'>here</a>");
-                mail.IsBodyHtml = true;
-                using (SmtpClient smtp = new SmtpClient("smtp.gmail.com", 587))
-                {
-                    string email = Environment.GetEnvironmentVariable("Email");
-                    smtp.UseDefaultCredentials = false;
-                    smtp.Credentials = new NetworkCredential(email, Environment.GetEnvironmentVariable("EmailPass"));
-                    smtp.EnableSsl = true;
-                    smtp.Send(mail);
-                }
-            }
+            using MailMessage mail = new MailMessage();
+            mail.From = new MailAddress("support@encryptionapiservices.com");
+            mail.To.Add(message.UserEmail);
+            mail.Subject = "Credit Card Changed - Encryption API Services";
+            mail.Body = "We noticed that you changed your credit card information recently. If this wasn't you we recommend changing your password " + String.Format("<a href='" + Environment.GetEnvironmentVariable("Domain") + "/#/forgot-password'>here</a>");
+            mail.IsBodyHtml = true;
+            SmtpClientSender.SendMailMessage(mail);
             this.Channel.BasicAck(deliveryTag: e.DeliveryTag, multiple: false);
         }
     }
